+++ src/lib.rs
@@ -57,7 +57,7 @@ pub mod gpu;
 pub mod wasm;
 
 pub use alphabet::{decode_char, encode_byte, encode_char};
-pub use alphabet::{Alphabet, AlphabetFns, DnaSequence, ExactDna, IupacDna};
+pub use alphabet::{Alphabet, AlphabetFns, DnaSequence, ExactDna, IupacDna, SymbolSet};
 pub use error::FmIndexError;
 pub use fm_index::bidir::BidirInterval;
 pub use fm_index::bidir_index::BidirFmIndex;